+++ config/software/gitlab-ctl.rb
@@ -63,7 +63,7 @@ done
 # As documented at http://smarden.org/runit/sv.8.html
 export SVWAIT=30
 
-#{install_dir}/embedded/bin/omnibus-ctl #{File.basename(install_dir)} #{install_dir}/embedded/service/omnibus-ctl $@
+#{install_dir}/embedded/bin/omnibus-ctl #{File.basename(install_dir)} '#{install_dir}/embedded/service/omnibus-ctl*' $@
        EOH
     end
   end