+++ files/gitlab-ctl-commands/lib/gitlab_ctl.rb
@@ -30,5 +30,7 @@ module GitlabCtl
     end
 
     class NodeError < StandardError; end
+
+    class PasswordMismatch < StandardError; end
   end
 end