+++ files/gitlab-ctl-commands/lib/postgresql.rb
@@ -1,0 +1,6 @@
+require_relative 'postgresql/pgpass'
+
+module GitlabCtl
+  class PostgreSQL
+  end
+end