+++ spec/support/matchers.rb
@@ -5,6 +5,8 @@ def create_ssh_key(path)
   ChefSpec::Matchers::ResourceMatcher.new(:ssh_keygen, :create, path)
 end
 
+ChefSpec.define_matcher :postgresql_user
+
 def create_postgresql_user(password)
   ChefSpec::Matchers::ResourceMatcher.new(:postgresql_user, :create, password)
 end